[FEATURE] Integrate Chef Cookbooks download in Vagrant file
Librarian and Bundle have been used as tools for managing
the Chef Cookbooks but it turns out it can be quite tricky
to install them on certain OS as Mac OS.
This change-set replaces the tool by a script which integrates
itself in the Vagrant file with the main benefit to drastically
reduces and simplify the installation procedure for the end user.

As a result, dependencies to librarian and bundler have been removed.

Change-Id: I7a875ca157892b45472ce158d6c4884f0435bba8
Reviewed-on: https://review.typo3.org/29465
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot